api/requirements.txt
django
djangorestframework
python-dotenv
SpeechRecognition
openai
django4-background-tasks